backend/requirements.txt
fastapi[standard]
selenium
python-dotenv
google-genai
pymongo
dnspython
passlib
python-jose
